refactor: 全栈代码质量审查与安全修复
安全修复:
- JWT密钥启动验证（长度≥32字符）、CORS来源外部化为配置属性
- JSON错误响应统一使用ObjectMapper替代字符串拼接，防止注入
- 移除docker-compose/alertmanager中的硬编码密码和占位符凭证
- 管理后台登录密码传输与后端BCrypt验证对齐
- GlobalExceptionHandler不再向客户端泄露堆栈和内部信息
- 输入验证增强：NaN/Infinity检查、参数范围校验、XSS防护

并发与可靠性修复:
- OrderService分布式锁泄漏修复（事务已提交后直接释放锁）
- pile.getPower() NPE防护，添加null检查与默认值
- 碳积分TOCTOU竞态修复，使用原子数据库更新
- MQ消费者幂等性保障：先处理后标记，Lua脚本原子重试计数
- WebSocket聊天系统消息使用昵称替代userId，防止信息泄露
- 401处理去重机制，防止短时间内多次重定向

前端修复:
- WebSocket重连指数退避策略、最大重连次数限制
- ECharts实例dispose防止内存泄漏、resize防抖优化
- 路由守卫JWT过期检查与角色验证
- Token过期时间从JWT payload提取，前端提前60秒缓冲
- 防重复提交机制、开放重定向防护

基础设施与配置:
- Redis缓存分层TTL策略（实时/准实时/稳定/静态四层）
- 缓存击穿/穿透/雪崩三重防护机制
- RabbitMQ死信队列完整处理链路
- 多数据源读写分离配置（Profile隔离）
- API限流拦截器（Redis Lua脚本实现滑动窗口）

代码清理:
- 删除误命名垃圾文件和agent生成的报告文档
- 清理注释代码块和未使用的TODO桩
- 统一错误响应格式和日志级别

+++ database/init.sql
@@ -510,7 +510,7 @@ CREATE TABLE `credit_product` (
     `points_required` INT NOT NULL COMMENT '所需积分',
     `stock` INT NOT NULL COMMENT '库存数量',
     `exchanged_count` INT DEFAULT 0 COMMENT '已兑换数量',
-    `original_price` DOUBLE DEFAULT NULL COMMENT '商品价值（元）',
+    `original_price` DECIMAL(8,2) DEFAULT NULL COMMENT '商品价值（元）',
     `status` TINYINT NOT NULL DEFAULT 1 COMMENT '状态：1上架 0下架',
     `limit_per_user` INT DEFAULT 0 COMMENT '每人限兑数量，0表示不限',
     `create_time` DATETIME NOT NULL DEFAULT CURRENT_TIMESTAMP COMMENT '创建时间',
